pyproject.toml
[tool.poetry]
name = "mailing-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
ipython = "^8.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
